requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99